+++ src/join.sol
@@ -24,19 +24,19 @@ pragma solidity ^0.8.20;
 // New deployments of this contract will need to include custom events (TO DO).
 
 interface GemLike {
-    function decimals() external view returns (uint256);
-    function transfer(address, uint256) external returns (bool);
-    function transferFrom(address, address, uint256) external returns (bool);
+  function decimals() external view returns (uint);
+  function transfer(address, uint) external returns (bool);
+  function transferFrom(address, address, uint) external returns (bool);
 }
 
 interface DSTokenLike {
-    function mint(address, uint256) external;
-    function burn(address, uint256) external;
+  function mint(address, uint) external;
+  function burn(address, uint) external;
 }
 
 interface VatLike {
-    function slip(bytes32, address, int256) external;
-    function move(address, address, uint256) external;
+  function slip(bytes32, address, int) external;
+  function move(address, address, uint) external;
 }
 
 /*
@@ -64,126 +64,126 @@ interface VatLike {
 */
 
 contract GemJoin {
-    // --- Auth ---
-    mapping(address => uint256) public wards;
+  // --- Auth ---
+  mapping(address => uint) public wards;
 
-    function rely(address usr) external auth {
-        wards[usr] = 1;
-        emit Rely(usr);
-    }
+  function rely(address usr) external auth {
+    wards[usr] = 1;
+    emit Rely(usr);
+  }
 
-    function deny(address usr) external auth {
-        wards[usr] = 0;
-        emit Deny(usr);
-    }
+  function deny(address usr) external auth {
+    wards[usr] = 0;
+    emit Deny(usr);
+  }
 
-    modifier auth() {
-        require(wards[msg.sender] == 1, "GemJoin/not-authorized");
-        _;
-    }
+  modifier auth() {
+    require(wards[msg.sender] == 1, "GemJoin/not-authorized");
+    _;
+  }
 
-    VatLike public vat; // CDP Engine
-    bytes32 public ilk; // Collateral Type
-    GemLike public gem;
-    uint256 public dec;
-    uint256 public live; // Active Flag
+  VatLike public vat; // CDP Engine
+  bytes32 public ilk; // Collateral Type
+  GemLike public gem;
+  uint public dec;
+  uint public live; // Active Flag
 
-    // Events
-    event Rely(address indexed usr);
-    event Deny(address indexed usr);
-    event Join(address indexed usr, uint256 wad);
-    event Exit(address indexed usr, uint256 wad);
-    event Cage();
+  // Events
+  event Rely(address indexed usr);
+  event Deny(address indexed usr);
+  event Join(address indexed usr, uint wad);
+  event Exit(address indexed usr, uint wad);
+  event Cage();
 
-    constructor(address vat_, bytes32 ilk_, address gem_) {
-        wards[msg.sender] = 1;
-        live = 1;
-        vat = VatLike(vat_);
-        ilk = ilk_;
-        gem = GemLike(gem_);
-        dec = gem.decimals();
-        emit Rely(msg.sender);
-    }
+  constructor(address vat_, bytes32 ilk_, address gem_) {
+    wards[msg.sender] = 1;
+    live = 1;
+    vat = VatLike(vat_);
+    ilk = ilk_;
+    gem = GemLike(gem_);
+    dec = gem.decimals();
+    emit Rely(msg.sender);
+  }
 
-    function cage() external auth {
-        live = 0;
-        emit Cage();
-    }
+  function cage() external auth {
+    live = 0;
+    emit Cage();
+  }
 
-    function join(address usr, uint256 wad) external {
-        require(live == 1, "GemJoin/not-live");
-        require(int256(wad) >= 0, "GemJoin/overflow");
-        vat.slip(ilk, usr, int256(wad));
-        require(gem.transferFrom(msg.sender, address(this), wad), "GemJoin/failed-transfer");
-        emit Join(usr, wad);
-    }
+  function join(address usr, uint wad) external {
+    require(live == 1, "GemJoin/not-live");
+    require(int(wad) >= 0, "GemJoin/overflow");
+    vat.slip(ilk, usr, int(wad));
+    require(gem.transferFrom(msg.sender, address(this), wad), "GemJoin/failed-transfer");
+    emit Join(usr, wad);
+  }
 
-    function exit(address usr, uint256 wad) external {
-        require(wad <= 2 ** 255, "GemJoin/overflow");
-        vat.slip(ilk, msg.sender, -int256(wad));
-        require(gem.transfer(usr, wad), "GemJoin/failed-transfer");
-        emit Exit(usr, wad);
-    }
+  function exit(address usr, uint wad) external {
+    require(wad <= 2 ** 255, "GemJoin/overflow");
+    vat.slip(ilk, msg.sender, -int(wad));
+    require(gem.transfer(usr, wad), "GemJoin/failed-transfer");
+    emit Exit(usr, wad);
+  }
 }
 
 contract DaiJoin {
-    // --- Auth ---
-    mapping(address => uint256) public wards;
+  // --- Auth ---
+  mapping(address => uint) public wards;
 
-    function rely(address usr) external auth {
-        wards[usr] = 1;
-        emit Rely(usr);
-    }
+  function rely(address usr) external auth {
+    wards[usr] = 1;
+    emit Rely(usr);
+  }
 
-    function deny(address usr) external auth {
-        wards[usr] = 0;
-        emit Deny(usr);
-    }
+  function deny(address usr) external auth {
+    wards[usr] = 0;
+    emit Deny(usr);
+  }
 
-    modifier auth() {
-        require(wards[msg.sender] == 1, "DaiJoin/not-authorized");
-        _;
-    }
+  modifier auth() {
+    require(wards[msg.sender] == 1, "DaiJoin/not-authorized");
+    _;
+  }
 
-    VatLike public vat; // CDP Engine
-    DSTokenLike public dai; // Stablecoin Token
-    uint256 public live; // Active Flag
+  VatLike public vat; // CDP Engine
+  DSTokenLike public dai; // Stablecoin Token
+  uint public live; // Active Flag
 
-    // Events
-    event Rely(address indexed usr);
-    event Deny(address indexed usr);
-    event Join(address indexed usr, uint256 wad);
-    event Exit(address indexed usr, uint256 wad);
-    event Cage();
+  // Events
+  event Rely(address indexed usr);
+  event Deny(address indexed usr);
+  event Join(address indexed usr, uint wad);
+  event Exit(address indexed usr, uint wad);
+  event Cage();
 
-    constructor(address vat_, address dai_) {
-        wards[msg.sender] = 1;
-        live = 1;
-        vat = VatLike(vat_);
-        dai = DSTokenLike(dai_);
-    }
+  constructor(address vat_, address dai_) {
+    wards[msg.sender] = 1;
+    live = 1;
+    vat = VatLike(vat_);
+    dai = DSTokenLike(dai_);
+  }
 
-    function cage() external auth {
-        live = 0;
-        emit Cage();
-    }
+  function cage() external auth {
+    live = 0;
+    emit Cage();
+  }
 
-    uint256 constant ONE = 10 ** 27;
+  uint constant ONE = 10 ** 27;
 
-    function mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
-        require(y == 0 || (z = x * y) / y == x);
-    }
+  function mul(uint x, uint y) internal pure returns (uint z) {
+    require(y == 0 || (z = x * y) / y == x);
+  }
 
-    function join(address usr, uint256 wad) external {
-        vat.move(address(this), usr, mul(ONE, wad));
-        dai.burn(msg.sender, wad);
-        emit Join(usr, wad);
-    }
+  function join(address usr, uint wad) external {
+    vat.move(address(this), usr, mul(ONE, wad));
+    dai.burn(msg.sender, wad);
+    emit Join(usr, wad);
+  }
 
-    function exit(address usr, uint256 wad) external {
-        require(live == 1, "DaiJoin/not-live");
-        vat.move(msg.sender, address(this), mul(ONE, wad));
-        dai.mint(usr, wad);
-        emit Exit(usr, wad);
-    }
+  function exit(address usr, uint wad) external {
+    require(live == 1, "DaiJoin/not-live");
+    vat.move(msg.sender, address(this), mul(ONE, wad));
+    dai.mint(usr, wad);
+    emit Exit(usr, wad);
+  }
 }
